+++ docs/changelog.rst
@@ -12,6 +12,8 @@ New features:
 Fixes:
 
 * :doc:`/plugins/mpdupdate`: Fix Python 3 compatibility. :bug:`2381`
+* :doc:`/plugins/replaygain`: Fix Python 3 compatibility in the ``bs1770gain``
+  backend. :bug:`2382`
 
 
 1.4.3 (January 9, 2017)